projects.js: Use project 'status' field
This introduces "status" field for projects. With this, users will
not have to use three different booleans, crowded, in_progress and
completed. Instead they can do,

"status" : "high involvement"
"status" : "in progress"
"status" : "completed"

The terminology is changed from crowded to high involvement as it
is more intuitive for the user.