app.js: Get mentors from Webservices
The webservices now has a table for mentors,
that stores the requests created by contributor
on the community website. After all checks are
green, the contributor is added a mentor for the
upcoming program based on the request. Therefore,
fetch all those contributors from webservices for
the upcoming GSoC program and append them to the
mentors list

Closes https://github.com/coala/projects/issues/728