tabs: Fetch and Display forms from webservices
The forms uploaded by developers from community
website, are being fetched from Webservices
where the request to add a new form is being
processed. If a valid request is created, the form
is added to the database. Therefore, we can fetch
those forms and display them on the projects website.

Closes https://github.com/coala/projects/issues/284