_projects: Use project title from work report
The commit adds the new title for all those completed
projects who got a new and a suitable title during the
program. As the student starts working on project, user
get to know "what will the be the most suitable title for
the project?" and uses that in the work report too. So,
add all such titles from reports to there corresponding
projects.
Also, this change will allow us to display the student work
report itself on the project popup.